+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Head
+# 2.0.0
 
 - Added: new version `eslint-plugin-jsx-a11y` and new rules from `eslint-plugin-jsx-a11y` plugin.
 - Added: new version `eslint-plugin-ava`.